创建订单

+++ src/main/java/com/fubuki/warship/model/pojo/Order.java
@@ -9,7 +9,7 @@ public class Order {
 
     private Long userId;
 
-    private Integer totalPrice;
+    private Long totalPrice;
 
     private String receiverName;
 
@@ -57,11 +57,11 @@ public class Order {
         this.userId = userId;
     }
 
-    public Integer getTotalPrice() {
+    public Long getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(Integer totalPrice) {
+    public void setTotalPrice(Long totalPrice) {
         this.totalPrice = totalPrice;
     }
